Remove IApplicationContext interface and everything that uses it.
The only purpose that this interface serves is to allow functionality that
depends on the physical path to the application. Since the Owin specification
doesn't expose this, we can't assume that we will have access to it, so don't
expose it here.

If needed, we can make it available for specific providers (e.g. the ASP.NET
provider) using extension methods.