Remove IsDependentOnRequest property from IFeatureSwitch.
This doesn't do what it's supposed to do. Case in point: when a
DateTimeFeatureSwitch times out, its Enabled state will change, say from off to
on. However, this change would be ignored by the FeatureSwitchboard, which would
then serve up a feature set that had been cached from before its state changed.
As a result, the feature in question would still be disabled even though it was
supposed to have been enabled.