Cleanly separate ProcessHeaders and ProcessBody in result processor
To do this, we need to allow our result processors to be stateful, so they need
to be truly transient. Unfortunately StructureMap has a rather esoteric
definition of "transient" when you're using nested containers -- they are scoped
to the nested container itself. Fortunately, however, it does provide an
AlwaysUnique lifecycle option which is transient in the sense that every other
IOC container in existence understands it.