Make feature switches and top-level IOC configuration mutually exclusive.
This restriction is to avoid confusion and potential bugs. Top-level IOC configuration (with `Setup.Application(c = >...)`) is always available to the application regardless of whether the feature is switched on or off. If we were to allow switchable features to specify application-level services, we could easily end up with code being switched on when it is believed to be switched off.