Shuffle the static files around. Add favicon.
In case you're wondering where the favicon and the logo come from, they are
both from a scene in the Hugo Award-winning xkcd webcomic "Time." [1]
Specifically, it's the scene where Cueball and Megan first meet the
Beanie Bunch. [2] "ᔪᕒᖚᐧ ᘛᔭᐤ" is the first that we see of the Beanish language.

One commentator [3] has hypothesised that it is pronounced "nalaka zaga"
and means something along the lines of "Where have you come from?" I have a
different theory. I would suggest that it is pronounced "Dolstagis."

[1] https://xkcd.com/1190
[2] http://geekwagon.net/projects/xkcd1190/?frame=2663
[3] https://beanishlang.wordpress.com/2014/04/03/back-to-the-corpus-i/