Common method for getting a streamwriter for the response.
This means that everything should work in the same way. It also fixes a bug
where the response locks up if an exception is thrown while trying to render
a view. This was due to the streamwriter disposing the underlying stream,
and then the exception handler subsequently attempting to write to it.